Fix test sessions not rolling back properly
So turned out tests sessions weren't actually rolling back, they just
weren't committed, because Transactional decorator was not using the
test session. Fixed it by properly replacing session in database.py.
Also added prepare_session function to use it both in the app and tests,
since in tests we need a new session maker binded to a transactional
connection to be able to rollback.