Go back to the table structure where client has address_id
I forgot that there will be a 'supplier' table which has address too. So
the structure where 'address' table has 'client_id' is not viable for
this. Now we go back to the structure where 'address' table has its own
id and the 'client' table has 'address_id'. That way I can now add
supplier with the address the same way.